ThemeChanger + Themes implemented
Created unique data asset "Theme" which holds the sprites of each layered restraunt asset.
Created ThemeChanger script. It holds all the themes can changes the sprites of the given restraunt pieces.

For now, there is a single CandyCane theme. The ThemeChanger script is on the BG in the restraunt and when space is pressed it applies the candycane theme to just the walls